Split of frect_t type, support for shooting.
The support for shooting isn't actually 'shooting', but adding
the ability to pull the trigger, more or less.

The frect_t type is going to be used a lot, as we'll be moving
objects around the screen with sub-pixel accuracy since the game
will run at a variable frame-rate.  Added the fpoint_t type while
I was at it.

Support for testing frect_t intersection is not yet supported,
but the stub is there in the header file.